requirements.txt
Django==1.11.8
django-autoslug==1.9.3
django-tinymce4-lite==1.7.0
jsmin==2.2.2
pytils==0.3
pytz==2017.3
